Don't use high baudrates when not available
On certain architectures (namely Sparc), the maximum baud rate exposed
by the kernel headers is B2000000. Therefore, the current libserial
code doesn't build for the Sparc and Sparc64 architectures due to
this.

In order to address this problem, this patch tests the value of
__MAX_BAUD. If it's higher than B2000000 then we assume we're on an
architecture that supports all baud rates up to B4000000. Otherwise,
we simply don't support the baud rates above B2000000.

Fixes build failures such as:

./SerialPort.h:88:24: error: 'B2500000' was not declared in this scope
         BAUD_2500000 = B2500000,

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>