Make the Python bindings build conditional
In some situations, building the Python bindings may not be desirable
or even possible, for example in the context of cross-compiled
embedded systems, Python may not even be available on the target
platform.

In order to address this, this commit adds a
--with-python/--without-python option to the configure script, which
allows to specify whether Python bindings should be built or not. Of
course, in order to preserve the current behavior, the Python bindings
are built by default (if neither --with-python nor --without-python is
passed).

This change would for example be useful for embedded Linux build
systems such as Buildroot (http://buildroot.org).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>