Refactored SerialStreamBuf so it uses an instance of SerialPort for its core functions.
This reduces a lot of code that was repeated between SerialStreamBuf and
SerialPort classes.

Also, marked several methods of both classes as "const" if they
represent quantities that should not change unless explicitly set by the
client. Methods such as IsDataAvailable() are not marked as "const" as
they might change without any action from clients of this class and,
hence, are not part of the constant state of an instance of these
classes.